chore: move lnc icon
