+++ src/components/icons/connectors/lncIcon.ts
